extract logic of component into composable
