+++ wls-gui-wahllokalsystem/src/components/ergebnisermittlung/OBW/stapelA/TheOBWStapelACard.vue
@@ -41,28 +41,15 @@
 </template>
 
 <script setup lang="ts">
-import type { ErgebnisAndWahlvorschlag } from "@/types/ergebnisermittlung/ErgebnisAndWahlvorschlag.ts";
-import type { Ergebnis } from "@/types/ergebnismeldung/Ergebnis.ts";
-
 import { computed, ref } from "vue";
 
 import BaseButtonSave from "@/components/common/buttons/BaseButtonSave.vue";
 import BaseRowObwStapelA from "@/components/ergebnisermittlung/OBW/stapelA/BaseRowOBWStapelA.vue";
-import { useLogging } from "@/composables/common/logging.ts";
-import { useErgebnisUtils } from "@/composables/ergebnismeldung/ergebnisUtils.ts";
+import { useOBWStapelAUtils } from "@/composables/ergebnisermittlung/obwStapelAUtils.ts";
 import { useErgebnismeldungStore } from "@/stores/ergebnismeldungStore.ts";
-import { useWahlvorschlaegeStore } from "@/stores/wahlvorschlaegeStore.ts";
 import { StapelArtEnum } from "@/types/ergebnismeldung/StapelArtEnum.ts";
 
-const {
-  sendErgebnisseByStapelArt,
-  getErgebnisseByWahlIdAndStapelartOrUndefined,
-} = useErgebnismeldungStore();
-const { getWahlvorschlagOrUndefinedByWahlIDWahlbezirkIDAndWahlvorschlagID } =
-  useWahlvorschlaegeStore();
-
-const { orderedByNumIndexWithNullAtEnd } = useErgebnisUtils();
-const { logWarn } = useLogging("TheOBWStapelACard");
+const { sendErgebnisseByStapelArt } = useErgebnismeldungStore();
 
 const STAPEL = StapelArtEnum.ObwA;
 
@@ -77,71 +64,14 @@ const props = defineProps({
   },
 });
 
-const isFormValid = ref<boolean | null>(null);
-
-const ergebnisseWithWahlvorschlag = computed<ErgebnisAndWahlvorschlag[]>(() => {
-  return createErgebnisseAndWahlvorschlaege();
-});
-
-const sumOfValidVotes = computed(() =>
-  ergebnisseWithWahlvorschlag.value
-    .map((item) => item.ergebnis)
-    .reduce((sum, ergebnis) => {
-      return sum + (ergebnis.ergebnis ?? 0);
-    }, 0)
+const { ergebnisseWithWahlvorschlag, sumOfValidVotes } = useOBWStapelAUtils(
+  computed(() => props.wahlID),
+  computed(() => props.wahlbezirkID)
 );
+
+const isFormValid = ref<boolean | null>(null);
 
 function onSaveClicked() {
   sendErgebnisseByStapelArt(props.wahlID, STAPEL);
-}
-
-function addWahlvorschlagForErgebnisIfExisting(
-  ergebnis: Ergebnis,
-  result: ErgebnisAndWahlvorschlag[]
-): void {
-  if (ergebnis.wahlvorschlagID) {
-    const wahlvorschlagForErgebnis =
-      getWahlvorschlagOrUndefinedByWahlIDWahlbezirkIDAndWahlvorschlagID(
-        props.wahlID,
-        props.wahlbezirkID,
-        ergebnis.wahlvorschlagID
-      );
-    if (wahlvorschlagForErgebnis) {
-      result.push({
-        ergebnis: ergebnis,
-        wahlvorschlag: wahlvorschlagForErgebnis,
-      });
-    } else {
-      logWarn(
-        `ergebnis wahlID=${props.wahlID}, wahlbezirkID=${props.wahlbezirkID}, wahlvorschlagID=${ergebnis.wahlvorschlagID} hat keinen Wahlvorschlag`
-      );
-    }
-  } else {
-    logWarn(
-      `ergebnis wahlID=${props.wahlID}, wahlbezirkID=${props.wahlbezirkID} ist ohne wahlvorschlagID`
-    );
-  }
-}
-
-function createErgebnisseAndWahlvorschlaege() {
-  const result: ErgebnisAndWahlvorschlag[] = [];
-
-  const ergebnisseForWahlAndStapel =
-    loadErgebnisseForWahlAndStapelOrderedByNumIndex();
-
-  ergebnisseForWahlAndStapel.forEach((ergebnis) => {
-    addWahlvorschlagForErgebnisIfExisting(ergebnis, result);
-  });
-
-  return result;
-}
-
-function loadErgebnisseForWahlAndStapelOrderedByNumIndex(): Ergebnis[] {
-  return (
-    getErgebnisseByWahlIdAndStapelartOrUndefined(
-      props.wahlID,
-      STAPEL
-    )?.ergebnisse.sort(orderedByNumIndexWithNullAtEnd) ?? []
-  );
 }
 </script>
